// frontend/src/app/core/api/models/todo-output.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TodoOutput {

  /**
   * Indica se a tarefa foi concluída
   */
  concluido?: boolean;

  /**
   * Data e hora de conclusão da tarefa
   */
  dataConclusao?: string;

  /**
   * Data e hora de criação da tarefa
   */
  dataCriacao?: string;

  /**
   * Descrição detalhada da tarefa
   */
  descricao?: string;

  /**
   * ID único da tarefa
   */
  id?: number;

  /**
   * Título da tarefa
   */
  titulo?: string;
}